Add a custom linker script to increase heap memory.
Due to the split memory map on the STM32L475, the default linker script
was not effectively using the SRAM2 memory. To make more room for the
heap, we can move the stack and statically allocated RAM into SRAM2.

Unfortunately, this also breaks the use of bit-banding in the DCFurs
LED matrix driver.