Use TIM8 and DMA2 to generate the matrix timing.
Free up some CPU time and increase the LED matrix scan speed by
using a circular DMA driven by a timer to produce the scanning
waveforms. This makes the API to drive the LED matrix the same as
the DC26 badge, except that we now expect TIM8 to be provided.