Increase color depth by adding more PWM steps.
For more visual awesomeness, we need more bit depth in the display, which
requires more PWM steps. We approximately try to achieve an RGB565. This
gives us 42 steps of red, 51 steps of green and 63 steps of blue after
color balancing.

To access the extra color space, a new set_pix_rgb() and set_pix_hue() have
been added. These take RGB24 or a hue/value pair which get converted to the
PWM steps.